feat: consolidate the cli and web server into one binary
`aura` and `aura-web-server` are the same product, sharing configuration
and release cadence, so shipping two executables added build and
packaging complexity without giving users a useful boundary.

Web-server startup moves out of `aura-web-server`'s `main` into a
`server` module the library exposes, and `aura` gains a `webserver`
subcommand whose trailing arguments pass straight to that parser. Both
entry points build the same router, timeouts and shutdown from one
`ServerArgs`, so every server flag and environment variable is
unchanged.

`aura-web-server` stays on as a shim that warns on stderr and delegates
to the same entry point, so existing service definitions keep working.
Docker, Compose, Helm, the install script, package metadata and the
in-repo docs move to `aura webserver`; the chart's `server.args` now
leads with `webserver`, so values overriding args must keep it first.

Fixes: #409
Fixes: #410
Ref: #353
