+++ crates/aura-web-server/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod a2a;
 pub mod handlers;
+pub mod server;
 pub mod session_store;
 pub mod streaming;
 pub mod types;
