+++ examples/complete/devops-assistant.toml
@@ -6,7 +6,7 @@
 #   export GITHUB_PERSONAL_ACCESS_TOKEN="ghp_..."
 #
 # Usage:
-#   CONFIG_PATH=examples/complete/devops-assistant.toml cargo run --bin aura-web-server
+#   CONFIG_PATH=examples/complete/devops-assistant.toml cargo run --bin aura -- webserver
 
 [mcp]
 sanitize_schemas = true
