+++ examples/minimal/bedrock.toml
@@ -10,7 +10,7 @@
 #
 # Usage:
 #   export AWS_REGION="us-east-1"
-#   CONFIG_PATH=examples/minimal/bedrock.toml cargo run --bin aura-web-server
+#   CONFIG_PATH=examples/minimal/bedrock.toml cargo run --bin aura -- webserver
 
 [agent]
 name = "Bedrock Assistant"
