+++ examples/minimal/openrouter.toml
@@ -10,7 +10,7 @@
 #
 # Usage:
 #   export OPENROUTER_API_KEY="..."
-#   CONFIG_PATH=examples/minimal/openrouter.toml cargo run --bin aura-web-server
+#   CONFIG_PATH=examples/minimal/openrouter.toml cargo run --bin aura -- webserver
 
 [agent]
 name = "OpenRouter Assistant"
